requirements.txt
pdfminer.six
iocextract
requests
ollama
python-dotenv
flask